requirements.txt
alembic==1.6.0
bcrypt==3.2.0
bidict==0.21.2
blinker==1.4
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
cryptography==3.4.7
dnspython==1.16.0
enum34==1.1.10
eventlet==0.31.0
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.10
Flask-Mail==0.9.1
flask-marshmallow==0.14.0
Flask-Migrate==2.7.0
Flask-SocketIO==5.1.0
Flask-SQLAlchemy==2.5.1
GlobalPayments.Api==1.0.5
greenlet==1.0.0
idna==2.10
importlib-metadata==3.4.0
itsdangerous==1.1.0
Jinja2==2.11.3
jsonpickle==2.0.0
Mako==1.1.4
MarkupSafe==1.1.1
marshmallow==3.11.1
marshmallow-sqlalchemy==0.25.0
pycparser==2.20
pygame==2.0.1
pyobjc==7.2
pyobjc-core==7.2
pyobjc-framework-Accessibility==7.2
pyobjc-framework-Accounts==7.2
pyobjc-framework-AddressBook==7.2
pyobjc-framework-AdServices==7.2
pyobjc-framework-AdSupport==7.2
pyobjc-framework-AppleScriptKit==7.2
pyobjc-framework-AppleScriptObjC==7.2
pyobjc-framework-ApplicationServices==7.2
pyobjc-framework-AppTrackingTransparency==7.2
pyobjc-framework-AuthenticationServices==7.2
pyobjc-framework-AutomaticAssessmentConfiguration==7.2
pyobjc-framework-Automator==7.2
pyobjc-framework-AVFoundation==7.2
pyobjc-framework-AVKit==7.2
pyobjc-framework-BusinessChat==7.2
pyobjc-framework-CalendarStore==7.2
pyobjc-framework-CallKit==7.2
pyobjc-framework-CFNetwork==7.2
pyobjc-framework-ClassKit==7.2
pyobjc-framework-CloudKit==7.2
pyobjc-framework-Cocoa==7.2
pyobjc-framework-Collaboration==7.2
pyobjc-framework-ColorSync==7.2
pyobjc-framework-Contacts==7.2
pyobjc-framework-ContactsUI==7.2
pyobjc-framework-CoreAudio==7.2
pyobjc-framework-CoreAudioKit==7.2
pyobjc-framework-CoreBluetooth==7.2
pyobjc-framework-CoreData==7.2
pyobjc-framework-CoreHaptics==7.2
pyobjc-framework-CoreLocation==7.2
pyobjc-framework-CoreMedia==7.2
pyobjc-framework-CoreMediaIO==7.2
pyobjc-framework-CoreMIDI==7.2
pyobjc-framework-CoreML==7.2
pyobjc-framework-CoreMotion==7.2
pyobjc-framework-CoreServices==7.2
pyobjc-framework-CoreSpotlight==7.2
pyobjc-framework-CoreText==7.2
pyobjc-framework-CoreWLAN==7.2
pyobjc-framework-CryptoTokenKit==7.2
pyobjc-framework-DeviceCheck==7.2
pyobjc-framework-DictionaryServices==7.2
pyobjc-framework-DiscRecording==7.2
pyobjc-framework-DiscRecordingUI==7.2
pyobjc-framework-DiskArbitration==7.2
pyobjc-framework-DVDPlayback==7.2
pyobjc-framework-EventKit==7.2
pyobjc-framework-ExceptionHandling==7.2
pyobjc-framework-ExecutionPolicy==7.2
pyobjc-framework-ExternalAccessory==7.2
pyobjc-framework-FileProvider==7.2
pyobjc-framework-FileProviderUI==7.2
pyobjc-framework-FinderSync==7.2
pyobjc-framework-FSEvents==7.2
pyobjc-framework-GameCenter==7.2
pyobjc-framework-GameController==7.2
pyobjc-framework-GameKit==7.2
pyobjc-framework-GameplayKit==7.2
pyobjc-framework-ImageCaptureCore==7.2
pyobjc-framework-IMServicePlugIn==7.2
pyobjc-framework-InputMethodKit==7.2
pyobjc-framework-InstallerPlugins==7.2
pyobjc-framework-InstantMessage==7.2
pyobjc-framework-Intents==7.2
pyobjc-framework-IOSurface==7.2
pyobjc-framework-iTunesLibrary==7.2
pyobjc-framework-KernelManagement==7.2
pyobjc-framework-LatentSemanticMapping==7.2
pyobjc-framework-LaunchServices==7.2
pyobjc-framework-libdispatch==7.2
pyobjc-framework-LinkPresentation==7.2
pyobjc-framework-LocalAuthentication==7.2
pyobjc-framework-MapKit==7.2
pyobjc-framework-MediaAccessibility==7.2
pyobjc-framework-MediaLibrary==7.2
pyobjc-framework-MediaPlayer==7.2
pyobjc-framework-MediaToolbox==7.2
pyobjc-framework-Metal==7.2
pyobjc-framework-MetalKit==7.2
pyobjc-framework-MetalPerformanceShaders==7.2
pyobjc-framework-MetalPerformanceShadersGraph==7.2
pyobjc-framework-MLCompute==7.2
pyobjc-framework-ModelIO==7.2
pyobjc-framework-MultipeerConnectivity==7.2
pyobjc-framework-NaturalLanguage==7.2
pyobjc-framework-NetFS==7.2
pyobjc-framework-Network==7.2
pyobjc-framework-NetworkExtension==7.2
pyobjc-framework-NotificationCenter==7.2
pyobjc-framework-OpenDirectory==7.2
pyobjc-framework-OSAKit==7.2
pyobjc-framework-OSLog==7.2
pyobjc-framework-PassKit==7.2
pyobjc-framework-PencilKit==7.2
pyobjc-framework-Photos==7.2
pyobjc-framework-PhotosUI==7.2
pyobjc-framework-PreferencePanes==7.2
pyobjc-framework-PubSub==7.2
pyobjc-framework-PushKit==7.2
pyobjc-framework-Quartz==7.2
pyobjc-framework-QuickLookThumbnailing==7.2
pyobjc-framework-ReplayKit==7.2
pyobjc-framework-SafariServices==7.2
pyobjc-framework-SceneKit==7.2
pyobjc-framework-ScreenSaver==7.2
pyobjc-framework-ScreenTime==7.2
pyobjc-framework-ScriptingBridge==7.2
pyobjc-framework-SearchKit==7.2
pyobjc-framework-Security==7.2
pyobjc-framework-SecurityFoundation==7.2
pyobjc-framework-SecurityInterface==7.2
pyobjc-framework-ServiceManagement==7.2
pyobjc-framework-Social==7.2
pyobjc-framework-SoundAnalysis==7.2
pyobjc-framework-Speech==7.2
pyobjc-framework-SpriteKit==7.2
pyobjc-framework-StoreKit==7.2
pyobjc-framework-SyncServices==7.2
pyobjc-framework-SystemConfiguration==7.2
pyobjc-framework-SystemExtensions==7.2
pyobjc-framework-UniformTypeIdentifiers==7.2
pyobjc-framework-UserNotifications==7.2
pyobjc-framework-UserNotificationsUI==7.2
pyobjc-framework-VideoSubscriberAccount==7.2
pyobjc-framework-VideoToolbox==7.2
pyobjc-framework-Virtualization==7.2
pyobjc-framework-Vision==7.2
pyobjc-framework-WebKit==7.2
pyOpenSSL==20.0.1
python-dateutil==2.8.1
python-dotenv==0.17.1
python-editor==1.0.4
python-engineio==4.1.0
python-socketio==5.3.0
pyttsx3==2.90
requests==2.25.1
six==1.16.0
SQLAlchemy==1.4.13
typing-extensions==3.10.0.0
urllib3==1.26.4
websocket-client==1.0.1
Werkzeug==1.0.1
xmltodict==0.12.0
zipp==3.4.0
